refactor(dh): move query to another location
